Wordle as separate util; new guess+solution lists
